Add AMD rs690 VID DID reporting and some minor cleanups.
Acked-by: Ronald G. Minnich <rminnich@gmail.com>
Acked-by: Marc Jones <marcj303@gmail.com>



git-svn-id: svn://svn.coreboot.org/coreboot/trunk@3786 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/southbridge/amd/rs690/chip.h
@@ -21,9 +21,9 @@
 #define RS690_CHIP_H
 
 /* Member variables are defined in Config.lb. */
-struct southbridge_amd_rs690_config 
+struct southbridge_amd_rs690_config
 {
-	unsigned long vga_rom_address;		/* The location that the VGA rom has been appened. */
+	u32 vga_rom_address;		/* The location that the VGA rom has been appened. */
 	u8 gpp_configuration;	/* The configuration of General Purpose Port, A/B/C/D/E. */
 	u8 port_enable;		/* Which port is enabled? GFX(2,3), GPP(4,5,6,7) */
 	u8 gfx_dev2_dev3;	/* for GFX Core initialization REFCLK_SEL */
